More magic price conversion stuff
