+++ Pipfile
@@ -20,7 +20,8 @@ dacite = "1.6.0"
 django-extensions = "3.1.3"
 typedate = "1.1.1"
 expiringdict = "1.2.1"
-
+stockholm = "0.4.4"
+forex-python = "1.6"
 
 [dev-packages]
 ipython = "*"
