+++ docker/env
@@ -21,5 +21,3 @@ STORE_NUMBER=+14158542097
 SIGNALD_PORT=15432
 ADMIN_USERNAME=admin
 ADMIN_PASSWORD=admin
-CURR_FREE_KEY=ec74bbf01ac0e01835d1
-CURR_PAID_KEY=pr_037ab72df8354f4f8d5e3c26062b3627
