+++ mobot/apps/common/currency.py
@@ -1,0 +1,8 @@
+from stockholm import BaseCurrency
+
+
+class MobileCoin(BaseCurrency):
+    ticker = "MOB"
+
+
+MOB = MobileCoin
