+++ mobot/apps/merchant_services/service.py
@@ -5,12 +5,17 @@ from mobot.apps.payment_service.service import PaymentService
 from mobot.apps.payment_service.models import Payment
 from mobot.apps.signald_client import Signal
 from validations import TwoModelValidation
+from phonenumbers import PhoneNumber
 
 
 ## Todo: Make this Async
-class MerchantService(Protocol):
+class DropService:
     payment_service: PaymentService
     signald_client: Signal
+
+    def add_customer(self, name: str, phone_number: str) -> Customer:
+        cust = Customer.objects.get_or_create(name=name, phone_number=phone_number)
+        return cust
 
     def find_single_drop_for_user(self, customer: Customer, drop: Drop, validations: List[TwoModelValidation[Customer, Drop]]) -> Optional[Drop]:
         found_drop = None
@@ -25,7 +30,7 @@ class MerchantService(Protocol):
 
     def find_drops_for_user(self, customer: Customer) -> List[Drop]:
         drops = Drop.objects.get()
-
+        
         pass
 
     def propose_drop_to_customer(self, customer: Customer, drop: Drop) -> List[Drop]: ...
@@ -42,6 +47,6 @@ if __name__ == "__main__":
     CUSTOMER_PHONE_NUMBER = "+44 7911 123456"
     cust = Customer.objects.get(phone_number=CUSTOMER_PHONE_NUMBER)
     drop = Drop.objects.get(id=ORIGINAL_AIRDROP_ID)
-    merchant_services = MerchantService()
+    drop_service = DropService()
 
     maybe_original_drop = merchant_services.find_single_drop_for_user(cust, drop)
