+++ mobot/settings/base.py
@@ -18,8 +18,6 @@ FULLSERVICE_ADDRESS = runtime_env('FULLSERVICE_ADDRESS')
 FULLSERVICE_PORT = runtime_env('FULLSERVICE_PORT')
 FULLSERVICE_URL = f"http://{FULLSERVICE_ADDRESS}:{FULLSERVICE_PORT}/wallet"
 PHONENUMBER_DB_FORMAT = "E164"
-CURR_FREE_KEY = runtime_env('CURR_FREE_KEY')
-CURR_PAID_KEY = runtime_env('CURR_PAID_KEY')
 
 FEE_PMOB = None
 ACCOUNT_ID = None
